RF + TST: Added back min_signal, testing with DTI data (not DSI data!).
DTI requires a b0 measurement. This did not exist in the data that was
previously used for testing. For that reason, the eigenvalues in these tests
were always exceedingly low (set by min_diffusivity, I believe). I believe that
it makes more sense to test the model fitting with DTI data. Single b-value and
including a b0 measurement.