TST + RF: Add test conditions for fractional streamline coordinates.
Also, fix up rounding in other places: in setting up the model, and in
vox2track.

Had to also reduce the match to the matlab version, but I am not sure
how close we want to stay to that, as we fix up things. At the very
least, the model RMSE is lower here than in the Dipy version, so the
model is a more accurate fit to the signal, for the same number of
parameters.