TST: Test for invariance of model_params to splitting of the data.
This looks like a bug to me: it shouldn't matter whether you fit the
entire data array or just one.

This happens because min_signal currently depends on the entire data array
that is passed, and that varies here depending on whether only one voxel
or the entire array is passed.