WIP: Improve the axonal water fraction estimation.
This patch set proposes to use the more accurate estimation of the axonal
water fraction (AWF) given by eq. (8) in Fieremans' paper. The current
implementation assumes that there is a direction for which the
intra-axonal space diffusivity Da,i for the i-th voxel is 0 for some
direction. This assumption may not be met in some circumstances, and hence
the expression of the current implementation (eq. (7)) is a lower bound
for the AWF.