Update gradients_spheres.py
There were some indicators of inline code that were incorrect.  I swtiched the order of a couple of things to make more natural (to my ear) sounding English.  Corrected a minor typo at the top.  Removed trailing blank lines in a blind attempt to fix the unrendered
```
.. admonition:: Example source code
```
that appears as a code block in
http://nipy.org/dipy/examples_built/gradients_spheres.html#example-gradients-spheres

Apologies in advance if I am off-base with any of the changes.