Modified the default output strategy to 'absolute' from 'append'. Details:
1) The default output strategy being used in the workflows was 'append'.
This was always creating the output files in the path where the first input
file is located. This behavior was updated to create the new output files
in the current working directory.

1.a) Two workflows have been tested after making this change to check if the
output file is being created in the expected location.

2) A collection of use cases to test the behavior of the workflows with respect to
where the output would be created under different circumstances for e.g.

2.a) When all the input files are in the same directory.
2.b) When input files are located in different directories.
2.c) When --mix_names is used and when it is not used.

The documentation of the test cases can be seen here,
https://github.com/parichit/dipy/tree/document_test_cases/dipy_test_cases