Workflow for visualizing the quality of registered image.
Note:
1) This workflow has been created by moving changes from an older branch ('affine_visual') because
there was the need to isolate all the visualization related changes to an individual workflow. Eventually,
these changes will be removed from the affine_visual branch because they will conflict with the affine_registration
workflow.

2) Some part of this code is being directly bought in from local files (which are not part of the repo) because
the code was being repeteadly for debugging and for improving the quality of the images.

First release of the workflow

1) Adding the function to create the mosaic from the slices. The function is different from the previous
commits (in the older branch 'affine_visual') because it also needs to read the afine matrix.

2) Made a minor addition to the image.py file to load the affine matrix for transforming the moved image to the
static image reference system.

3) Added a static method for checking the dimensions of the static and the moved image.