Added the beltrami_affine function.
This function computes the Beltrami and Levi-Civita incremnts to update the
Iwasawa coordinates when the affine metric is chosen, the computations are
more complex than those of the euclidean metric.

Note: When the euclidean metric is chosen, the beltrami_euclidean function
is called instead. This function acts directly on the diffusion tensor
coordinates (no need for Iwasawa conversion) and there is no need to compute
the Levi-Civita terms.