Added the fidelity_affine function.
This function computes the Fidelity increments to update the Iwasawa
coordinates when the affine metric is chosen over the euclidean. This version
of the fidelity function requires the symbolic derivatives of the diffusion
tensor with respect to each Iwasawa coordinate, which are computed with the
get_qDXiq functions (added in the previous commit).