Added the beltrami_main.py file to dipy/reconst.
This file calls the functions from reconst/beltrami.py to implemnt the main
gradient descent routine and estimate the Free Water fraction, also, it deals
with pre-processing of the raw data and initialization of the diffusion
tensor and free water fraction.

Note: Eventually the fucntions from this file should be integrated into the
reconst/fwdti.py module.