BF: Fix NumPy warning when creating arrays from ragged sequences
Fix NumPy warning when creating arrays from ragged sequences in streamline test.

Fixes

```
VisibleDeprecationWarning: Creating an ndarray from ragged nested sequences
(which is a list-or-tuple of lists-or-tuples-or ndarrays with different
lengths or shapes) is deprecated. If you meant to do this, you must specify
'dtype=object' when creating the ndarray
orig_streamlines_world = list(np.subtract(new_streamlines_world, disps))
```

triggered in
https://travis-ci.org/github/dipy/dipy/jobs/676101427#L5569
and
https://travis-ci.org/github/dipy/dipy/jobs/676764173#L5561

for example.

Rationale:
https://numpy.org/neps/nep-0034-infer-dtype-is-object.html