ENH: Return unique invalid streamline removal indices
Avoid returning length-fold indices for the streamline indices to be
removed. Streamlines are uniquely identified by a single index in the
offset list.

For each streamline that needs to be removed, `ic_offsets_indices`
contains the range of values starting from the offset value at issue
until the offset value plus the length of the streamline. Hence, the
bisect function is keeping as many insertion points/values as the
corresponding length for every streamline offset to be removed.

Add the necessary tests to ensure that the expected indices are being
kept/removed.