TEST: Remove legacy `nose`-related dead testing code
Remove legacy `nose`-related dead testing code: `run_module_suite` is
a `NumPy` testing method to run a test module that uses the `nose`
testing framework under the hood. DIPY switched to `pytest` as its
testing framework in a series of commits merged on Jan 26, 2019, and
with the 51234a4 merge commit.