TEST: Adding random generator with seed to icm tests
This commit adds seeds to test_square_iter and test_icm_square to
stabilize the tests. square_gauss and square_1, which were originally
global variables are now loaded through a function which has a numpy
random generator inside with fixed seed. The variables were wrapped in a
function because 1. there is no need for the variables to be global 2.
this complies with a future PR that will add decorators on tests that
require seed values.