TEST: Stabilizing tensorflow training tests
This is a commit to stabilize tests that require specific
loss/accuracy as a training result by increasing the number
of training epochs. With the change in the numpy random system,
it is possible that it disrupted the initialization of the model,
reducing the expected accuracy/loss in the assertions.