RF: Moving to numpy.random.Generator from numpy.random
This commit is to move from numpy.random to numpy.random.Generator
as numpy suggests. All tests that require random numbers are now using
set_random_number_generator decorator from dipy.testing. Some functions
were changed to accept rng (or replaced seed with rng) as an argument so
that this complies with the changes made to the test. Randomstates were
changed to Generators as well. Some seeds were modified for the test to
work. test_mrf.py's global parameters are now all wrapped in a function.