BF: Removing Error/Warning from Tensorflow 2.16
Keras3, a new default for Tensorflow, deprecates the orthogonal option when initializing layers, which is addressed by directly calling the class. It also deprecates the ability to control number of processes and workers when running the model, which the PR removes. Both were previously throwing errors in CNN1Denoising. Finally, it also has a change of behavior in how the input dictionary's tensors should be named, throwing warning in EVACPlus, which is also taken care of in this PR.