NF: Adding correct_mask to median_otsu
This PR aims to add a function that removes holes or islands of a binary
mask to median_otsu.

The function was already being used in
dipy/nn/utils.py, so it has been moved to dipy/segment/utils.py, as it
makes more sense for mask correction to be in segmentation module.

Slice-wise hole removal was excluded from the function, since it can
cause huge false positives if there is a lack of axial slices.

Finally, the function's name has been changed to
remove_holes_and_islands, to be more explicit about the functionality.

Related parameter name and its descriptions have been changed as well.