STYLE: Set `stacklevel` argument explicitly to warning messages
Set `stacklevel` argument explicitly to warning messages in
`testing.tests.test_testing.py` module. Set its value to 1, as opposed
to the `ruff` recommendation of using 2 [1], since otherwise the tests
do not pass: the `__warningregistry__` seems not to be created (and
cannot be queried) otherwise. This is related to the fact that the
`stacklevel` option to warnings causes the attribute to be set for the
module the warning was issued "in the name of", not necessarily the one
where `warn()` was called, e.g. if we specify a value larger than 1.

Fixes:
```
dipy/testing/tests/test_testing.py:52:9:
 B028 No explicit `stacklevel` keyword argument found
dipy/testing/tests/test_testing.py:56:9:
 B028 No explicit `stacklevel` keyword argument found
dipy/testing/tests/test_testing.py:60:9:
 B028 No explicit `stacklevel` keyword argument found
dipy/testing/tests/test_testing.py:65:9:
 B028 No explicit `stacklevel` keyword argument found
```

Remove the corresponding rule from the `ruff` per-file ignore list.

[1] https://docs.astral.sh/ruff/rules/no-explicit-stacklevel/