STYLE: Make module level import be at top of file
Make module level import be at the top of the file.

Fixes:
```
dipy/align/__init__.py:27:1: E402 Module level import not at top of file
```

Remove the corresponding rule from the `ruff` per-file ignore list.

Relocate the `np.float32` aliasing to an `align.utils.py` module to
avoid a circular import error.

Keep the list of semantically public names of the module (`__all__`) at
the `__init__.py` file, and avoid duplicating the list in the
`_public.py` file.

Relocate the `align.Bunch` class and the `align.VerbosityLevels` enum to
`utils` since they are general enough so as to be potentially used
elsewhere.