RF: Do not use `np.any` for checking optional array parameters
Do not use `np.any` for checking optional array parameters: prior to
NumPy 2.0 `np.any` would not return booleans for object `dtype` arrays,
and thus, it would return `None` for an input that is `None` and the
comparison against `None` would hold; however, starting NumPy 2.0 it
will return `False` and the comparison `False is None` would be `False`,
and hence the weight arrays would remain as `None`.

This fix is in preparation for full NumPy 2.0 compatibility.

Fixes:
```
>       imagec = abs(_fft.ifft2(_fft.fftshift(C1) * G1 + _fft.fftshift(C0) * G0))
E       TypeError: unsupported operand type(s) for *: 'complex' and 'NoneType'

../venv/lib/python3.12/site-packages/dipy/denoise/gibbs.py:237: TypeError
```

and similar errors.

Reported for example in:
https://github.com/dipy/dipy/actions/runs/9606192502/job/26495290276#step:10:6715

Documemtation:
https://numpy.org/doc/2.0/reference/generated/numpy.any.html