DOC: Use math mode for SH basis cases documentation
Use math mode for SH basis cases documentation: improves readability
when the documentation is rendered. Mark the corresponding docstrings as
raw strings.

Make the multiplying factor be before the `Re` and `Im` factors for the
sake of readability.

Use curly brackets to enclose the exponent of the basis so that the
LaTeX expression is correctly rendered (i.e. $Y^{|m|}_l$ instead of
$Y^|m|_l$, where only the first vertical bar would be raised).

Take advantage of the commit to use LaTeX markup when the SH order and
degree symbols are mentioned in docstrings.