DOC: Enclose star-containing docstring text with double backticks
Enclose star-containing docstring text with double backticks: stars that
are not followed by a whitespace are used in pairs with a matching
ending to italicize a string. In this case, stars appear under the form
of `*args` and `**kwargs` as an example of a function signature, so no
listing or italicizing was intended.

Fixes:
```
/dipy/reconst/dki_micro.py:docstring of dipy.reconst.dki_micro.KurtosisMicrostructureModel.__init__:13:
 WARNING: Inline emphasis start-string without end-string.
```

raised for example in:
https://github.com/dipy/dipy/actions/runs/10310154443/job/28541321104#step:5:1067