DOC: Use the appropriate indentation in docstrings
Use the appropriate indentation in docstrings and/or make it more
readable.

For multi-line item lists, make all lines have the same indentation.

Take advantage to remove default parameter values from the modified
docstrings to reduce the maintenance burden and avoids the risk of
inconsistencies if the default value is changed in the method
signature but not the parameter docstring.