DOC: Enclose inline literal in double backticks
Enclose inline literal in double backticks: add missing ending double
backticks.

Take advantage of the commit to remove its default value: reduces the
maintenance burden and avoids the risk of inconsistencies if the default
value is changed in the method signature but not the parameter
docstring.

Fixes:
```
/dipy/segment/bundles.py:docstring of dipy.segment.bundles.RecoBundles.evaluate_results:27:
 WARNING: Inline literal start-string without end-string.
```

raised for example in:
https://github.com/dipy/dipy/actions/runs/10472634043/job/29002561651#step:5:1063