STYLE: Improve first interaction text layout
Improve first interaction text layout: split the text into more
paragraphs, make them shorter if possible, and do not introduce line
breaks manually within paragraphs.

The line breaks introduced manually in the middle of the sentences to
limit to a given line length in the previous version made the text to
be broke accordingly. This resulted in a weird layout.

Make the exclamation marks more visible by introducing a whitespace
before them.