RF: Editing tranform_img in dipy.nn.utils
The transform_img function did not have an option to make an image isotropic and have a fixed size. This adds a way of doing this by padding to a multiple of that shape and resizing.
Recover_img has been changed to compensate for the difference as well. The parameters returned by transform_img, which was only used in recover_img, was simplified to make the code cleaner.
Deepn4 strangely had different results between the old transform_img and new, but the output when checked did not have meaningful differences. Hence, test output was edited to make the test functional.