BF: Make `StoppingCriterion` reproducible for multi-thread execution
Make `StoppingCriterion` reproducible for multi-thread execution:
instantiate the resulting interpolation array at each call.

In single-thread mode the maps created by the criterion are
reproducible; however, in multi-thread execution the
`self.interp_out_view` is modified in parallel by the multiple threads,
and since it is instantiated only once, the maps are not reproducible
across runs.

This effect is only observed in criterion objects that use the
`self.interp_out_view` array: i.e. it is not observed in the
`BinaryStoppingCriterion` objects.