RF: Remove unreachable code in tractogram saving function
Remove unreachable code in tractogram saving function: any statement
after a `raise` statement is unreachable.

More generally, make the function not return anything: it does not make
sense to only be able to return a value that can only be `True` or
`None` (if an exception is raised) in this case. If no exception is
raised, the user can expected with confidence that the file was saved
correctly.

Introduced in commits e0d3c9c and e1ab89c.

For a comparison, `save_surface` does not return anything.