RF: Editing tranform_img in dipy.nn.utils (#3542)
RF: The transform_img function did not have an option to make an image isotropic and have a fixed size. This adds a way of doing this by padding to a multiple of that shape and resizing. Recover_img has been changed to compensate for the difference as well. The parameters returned by transform_img, which was only used in recover_img, was simplified to make the code cleaner. Test data for evac and deepn4 has been changed to match the behavior. Reslice was added a final shape parameter to enforce a shape after reslicing voxel value.