Adding synthseg pytorch model
This PR adds a pytorch version of the synthseg model.
Major changes include synthseg in nn, tests and data connected to the model, and utils.py in segment to mimic a postprocessing behavior
that is in the original model pipeline. A bug in transform_img has been taken care of as well. Some model tests had to be changed
accordingly.