Adding synthseg pytorch model
This PR adds a pytorch version of the synthseg model.
Major changes include synthseg in nn, tests and data connected to the model, and utils.py in segment to mimic a postprocessing behavior
that is in the original model pipeline. A bug in transform_img has been taken care of as well. Some model tests had to be changed
accordingly. Added in ClassifyTissue workflow. Force flag was added in workflow tests so that it will run regardless of temporary file's existance.
Minor default output file name issue was fixed in nn workflow.